backend/requirements.txt
python-dotenv==0.20.0
Flask==2.0.3
Flask-Cors==3.0.10
numpy==1.22.2
pandas==1.4.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.31